Separate the concept of the README from the tutorial content
Since this repo has other files, I think it makes sense to have a README
for the repo that is distinct from the tutorial content. That gives us a
place describe the other files, talk about ourselves, and hang a license

I put the entire content of the tutorial in tutorial.md, but the intro
might need some editing since I copied and edited some of that.