Modify readme and intro disclaimers
I don't think it's useful to have "Disclaimers" as a main section in the
readme, though I do think that some version of that content is good. I
also think we only need to ask for feedback in one of the 2 places, and
the readme seems more appropriate to me.